Introduce a flexible impl strategy for TimersIdlingResource, paving way for RN 62
